requirements.txt
flask>=3.0.0
anthropic>=0.40.0
google-genai>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
gunicorn>=22.0.0
